requirements.txt
boto3==1.20.0
psycopg2-binary==2.9.3
pandas==1.3.4
